+++ .changeset/happy-cameras-travel.md
@@ -1,0 +1,5 @@
+---
+'datezen': minor
+---
+
+feat(format): add localization for weekday and month tokens
