+++ scripts/strip-plugin.js
